refactor: web -> browser
