+++ lib/browser.test.js
@@ -1,4 +1,4 @@
-import { debouncePayloadSync } from './web';
+import { debouncePayloadSync } from './browser';
 
 describe('debouncePayloadSync', () => {
   let storage;
